Move task related code into own folder
