+++ crates/builder/src/lib.rs
@@ -4,7 +4,7 @@ mod generate;
 mod install;
 mod postbuild;
 mod prebuild;
-mod task;
+mod tasks;
 mod util;
 
 pub use config::{CmdArgs, Commands, Config};
